pankit-bot-backend/requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
sentence-transformers
httpx
